+++ io/serial.h
@@ -1,9 +1,12 @@
 #ifndef SERIAL_H
 #define SERIAL_H
+
+#include "../lib/stdint.h"
+
 /* All the I/O ports are calculated relative to the data port. This is because
-    * all serial ports (COM1, COM2, COM3, COM4) have their ports in the same
-    * order, but they start at different values.
-    */
+ * all serial ports (COM1, COM2, COM3, COM4) have their ports in the same
+ * order, but they start at different values.
+ */
 
 #define SERIAL_COM1_BASE                0x3F8      /* COM1 base port */
 
@@ -16,9 +19,9 @@
 /* The I/O port commands */
 
 /* SERIAL_LINE_ENABLE_DLAB:
-    * Tells the serial port to expect first the highest 8 bits on the data port,
-    * then the lowest 8 bits will follow
-    */
+ * Tells the serial port to expect first the highest 8 bits on the data port,
+ * then the lowest 8 bits will follow
+ */
 #define SERIAL_LINE_ENABLE_DLAB         0x80
 
 /** serial_configure_baud_rate:
@@ -29,7 +32,7 @@
 *  @param com      The COM port to configure
 *  @param divisor  The divisor
 */
-void serial_configure_baud_rate(unsigned short com, unsigned short divisor);
+void serial_configure_baud_rate(uint16_t com, uint16_t divisor);
 
 /** serial_configure_line:
 *  Configures the line of the given serial port. The port is set to have a
@@ -38,7 +41,7 @@ void serial_configure_baud_rate(unsigned short com, unsigned short divisor);
 *
 *  @param com  The serial port to configure
 */
-void serial_configure_line(unsigned short com);
+void serial_configure_line(uint16_t com);
 
 /** serial_write
 * Sends a buffer of length len to the specified serial port.
@@ -47,8 +50,8 @@ void serial_configure_line(unsigned short com);
 * @param len - The length of the buffer
 * @param com - The COM port to use
 */
-int serial_write(char * buf, unsigned int len, unsigned short com);
+uint32_t serial_write(char * buf, size_t len, uint16_t com);
 
-int serial_is_transmit_fifo_empty(unsigned int com);
+uint32_t serial_is_transmit_fifo_empty(uint32_t com);
 
 #endif /* SERIAL_H */
