+++ io/write.h
@@ -1,6 +1,6 @@
 #ifndef WRITE_H
 #define WRITE_H
-
-int write(char * buf, unsigned int len);
+#include "../lib/stdint.h"
+int write(char * buf, size_t len);
 
 #endif /* WRITE_H */
